+++ ferry/database/diff_db.py
@@ -18,6 +18,14 @@ from ferry.transform import transform
 
 queries_dir = Path(__file__).parent / "queries"
 
+primary_keys = {
+        "flags" : ["flag_id"],
+        "course_flags" : ["course_id"],
+        "professors" : ["professor_id"],
+        "course_professors" : ["course_id"],
+        "courses" : ["course_id"],
+        "listings" : ["listing_id"],
+    }
 
 def get_dfs(database_connect_string: str):
     db = Database(database_connect_string)
@@ -96,16 +104,8 @@ def check_change(row, table_name):
 
 def generate_diff(tables_old: dict[str, pd.DataFrame],
                     tables_new: dict[str, pd.DataFrame], output_dir:str):
-    
-    primary_keys = {
-        "flags" : ["flag_id"],
-        "course_flags" : ["course_id"],
-        "professors" : ["professor_id"],
-        "course_professors" : ["course_id"],
-        "courses" : ["course_id"],
-        "listings" : ["listing_id"],
-    }
 
+    diff_dict = {}
 
     for table_name in primary_keys.keys():
         if table_name not in tables_new.keys() or table_name not in tables_old.keys():
@@ -113,7 +113,7 @@ def generate_diff(tables_old: dict[str, pd.DataFrame],
         
         print(f"Computing diff for table {table_name} ...", end=" ")
         
-        output_file_path = Path(output_dir) / (table_name + ".txt")
+        output_file_path = Path(output_dir) / (table_name + ".md")
         
 
         with open(output_file_path, "w+") as file:
@@ -121,11 +121,6 @@ def generate_diff(tables_old: dict[str, pd.DataFrame],
             old_df = tables_old[table_name]
             new_df = tables_new[table_name]
             
-            ## if want to test with differences between old and new df
-            # if (table_name == "courses"):
-            #     # temporary -- can be created by uncommenting where it says just testing
-            #     old_df = pd.read_csv("/workspaces/ferry/new_df.csv")
-
             # TODO - better way to do this?
             pk = primary_keys[table_name][0]
 
@@ -133,26 +128,18 @@ def generate_diff(tables_old: dict[str, pd.DataFrame],
             
             # check for rows that are in old df but not in new df
             # based on primary key
+            file.write("## Rows missing in new table: \n")
+
             missing_rows = old_df[~old_df[pk].isin(new_df[pk])]
             if not missing_rows.empty:
-                file.write(f"Rows missing in new table: {missing_rows.to_csv()}\n")
+                file.write(f"{missing_rows.to_csv()}\n")
             
-            # check for rows that have been deleted
-            deleted_rows = new_df[~new_df[pk].isin(old_df[pk])]
-            if not deleted_rows.empty:
-                file.write(f"Deleted rows in new table: {deleted_rows.to_csv()}\n")
+            file.write("## Added rows in new table: \n")
+            # check for rows that have been added
+            added_rows = new_df[~new_df[pk].isin(old_df[pk])]
+            if not added_rows.empty:
+                file.write(f"{added_rows.to_csv()}\n")
 
-            # check for row
-
-            # just testing
-            # old_df.to_csv("old_df.csv", index=False)
-            # new_df.to_csv("new_df.csv", index=False)
-            
-            # check for rows that have changed
-            # changed_rows = old_df[((~old_df.isna()) & (~new_df.isna()) & (old_df != new_df)).any(axis=1)]
-            
-            # if not changed_rows.empty:
-            #     file.write(f"Changed rows in new table: {changed_rows}\n")
             if table_name == "course_flags":
                 old_df = old_df.groupby("course_id")["flag_id"].apply(frozenset)
                 new_df = new_df.groupby("course_id")["flag_id"].apply(frozenset)
@@ -166,143 +153,21 @@ def generate_diff(tables_old: dict[str, pd.DataFrame],
             
             changed_rows = merged_df[merged_df.apply(check_change, args=(table_name,), axis=1)]
             
+            file.write("## Changed rows in new table: \n")
+
             if not changed_rows.empty:
-                file.write(f"Changed rows in new table: {changed_rows.to_csv()}\n")
-        
+                file.write(f"{changed_rows.to_csv()}\n")
+
+            diff_dict[table_name] = {
+                "missing_rows": missing_rows,
+                "added_rows": added_rows,
+                "changed_rows": changed_rows
+            }
+
         print("✔")
 
+    return diff_dict
 
 tables_old = get_dfs("postgresql://postgres:postgres@db:5432/postgres")
 tables_new = transform(data_dir=Path("/workspaces/ferry/data"))
 generate_diff(tables_old, tables_new, "/workspaces/ferry/diff")
-
-def sync_db(tables: dict[str, pd.DataFrame], database_connect_string: str):
-    db = Database(database_connect_string)
-
-    # sorted tables in the database
-    db_meta = MetaData()
-    db_meta.reflect(bind=db.Engine)
-
-    # First step: existing -> old
-    print("\nMoving existing tables...")
-
-    conn = db.Engine.connect()
-    inspector = inspect(db.Engine)
-    
-    # TODO: remove this; currently we have a deadlock issue when executing on prod DB
-    conn.execution_options(isolation_level="AUTOCOMMIT")
-    replace = conn.begin()
-    conn.execute(text("SET CONSTRAINTS ALL DEFERRED;"))
-    for table in Base.metadata.sorted_tables:
-        logging.debug(f"Updating table {table}")
-        # If table doesn't exist, skip
-        if table.name not in db_meta.tables:
-            logging.debug(f"Table {table} does not exist in database.")
-            continue
-        # remove the old table if it is present before
-        conn.execute(text(f"DROP TABLE IF EXISTS {table}_old CASCADE;"))
-        for index in inspector.get_indexes(table.name):
-            index_name = index["name"]
-            if not index_name:
-                continue
-            conn.execute(text(f"ALTER INDEX {index_name} RENAME TO {index_name}_old"))
-
-        for constraint in [
-            inspector.get_pk_constraint(table.name),
-            *inspector.get_foreign_keys(table.name),
-            *inspector.get_unique_constraints(table.name),
-        ]:
-            name = constraint["name"]
-            if not name:
-                continue
-            conn.execute(
-                text(f"ALTER TABLE {table} RENAME CONSTRAINT {name} TO {name}_old")
-            )
-        # rename current main table to _old
-        # (keep the old tables instead of dropping them
-        # so we can rollback in case of errors)
-        # Note that this is done after we've retrieved the indexes and constraints
-        conn.execute(text(f'ALTER TABLE IF EXISTS "{table}" RENAME TO {table}_old;'))
-
-    replace.commit()
-
-    print("\033[F", end="")
-    print("Moving existing tables... ✔")
-
-    # Second step: stage new tables
-    print("\nAdding new staging tables...")
-    # TODO this should probably be done within one transaction
-    conn = db.Engine.raw_connection()
-    Base.metadata.create_all(db.Engine)
-    for table in Base.metadata.sorted_tables:
-        if table.name not in tables:
-            raise ValueError(
-                f"{table.name} defined in Base metadata, but there is no data for it."
-            )
-        # create in-memory buffer for DataFrame
-        buffer = StringIO()
-        # TODO is this really needed?
-        tables[table.name] = tables[table.name].replace({r"\r": ""}, regex=True)
-        tables[table.name].to_csv(
-            buffer,
-            index_label="id",
-            header=False,
-            index=False,
-            sep="\t",
-            quoting=csv.QUOTE_NONE,
-            escapechar="\\",
-            na_rep="NULL",
-        )
-
-        buffer.seek(0)
-        cursor = conn.cursor()
-
-        try:
-            cursor.copy_from(
-                buffer,
-                table.name,
-                columns=tables[table.name].columns,
-                sep="\t",
-                null="NULL",
-            )
-        except Exception as error:
-            conn.rollback()
-            cursor.close()
-            raise error
-
-        # print(f"Successfully copied {table_name}")
-        cursor.close()
-    conn.commit()
-
-    print("\033[F", end="")
-    print("Adding new staging tables... ✔")
-
-    # Last: drop the _old tables
-    print("\nDeleting temporary old tables...")
-
-    conn = db.Engine.connect()
-    delete = conn.begin()
-    conn.execute(text("SET CONSTRAINTS ALL DEFERRED;"))
-    for table in Base.metadata.sorted_tables:
-        logging.debug(f"Dropping table {table}_old")
-        conn.execute(text(f"DROP TABLE IF EXISTS {table}_old CASCADE;"))
-    delete.commit()
-
-    print("\033[F", end="")
-    print("Deleting temporary old tables... ✔")
-
-    print("\nReindexing...")
-    conn.execution_options(isolation_level="AUTOCOMMIT")
-    conn.execute(text("REINDEX DATABASE postgres;"))
-    print("\033[F", end="")
-    print("Reindexing... ✔")
-
-    # Print row counts for each table.
-    print("\n[Table Statistics]")
-    with database.session_scope(db.Session) as db_session:
-        with open(queries_dir / "table_sizes.sql") as file:
-            SUMMARY_SQL = file.read()
-
-        result = db_session.execute(text(SUMMARY_SQL))
-        for table_counts in result:
-            print(f"{table_counts[1]:>25} - {table_counts[2]:6} rows")
